+++ src/test/python/build_system.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+class build_system(Enum):
+	MAVEN = 1
+	GRADLE = 2
